code refactor, syntax change
