+++ italy_geopop/__init__.py
@@ -1,3 +1,4 @@
-from .geopop import ItalyGeopopDataFrame
-from .pandas_extension import activate as pandas_activate
-from .pandas_extension import activate_context as pandas_activate_context
+from .__version__ import __version__
+
+from . import pandas_extension
+from . import geopop
